Refactor nav control architecture
